quest 04

+++ sqrt.go
@@ -3,5 +3,5 @@ package piscine
 import "math"
 
 func Sqrt(int nb) int {
-	return math.Sqrt(nb)
+	return math.Sqrt(float64(nb))
 }
